Seamless CSV report
If resource is not changed during time period we have only
one record in DB. But in CSV time series report we should have
rows on each date for resource. For such case we are duplicating
not changed row in the CSV report. But we should remove from
resource data 'added', 'modified', 'removed' in case of duplication
of the resource row.
We are duplicating only resource rows only until masternode had sent
stats. For indication of sending stat InstallationStructure creation_date
and modification_date are used.

Implements: blueprint openstack-workload-statistics
Change-Id: Ide21e17d5ebf8359fe2913a5458c9c8d499b8da5