Data format of workload resource_data changed
Fuel-stats part of implementation.
In case of cluster reset we can lose information about adding and
removing workloads. For prevention format of resource_data is
changed. Now we have lists of dicts in 'added', 'remove', 'modified'
and 'current'. On cluster reset we can have several records in the
'added' with different time of adding. For 'removed' we add full
data only if it doesn't already added. In other case we add only
id with removing time.

Change-Id: I2d57f53fcb1489f177772875464056ef301b16a6
Closes-Bug: #1427757
Implements: blueprint openstack-workload-statistics