Fuel version from cluster used in OSWL reports
Wrong source for Fuel version was used in the OSWL report.
We used the version of Fuel installed on master node, instead of
cluster version. Thus in case of upgrade master node we had
different fuel versions in the clusters and OSWL reports for
the same master node uid.

Now we are using Fuel version from OSWL.version_info or from
the cluster data. Fuel version from cluster is used also in the
plugins report.

Change-Id: I114beae261686fb6e4d77504dd102b722a43e611
Closes-Bug: #1547565