Cleaning test DB without Alembic migrations
On the Fuel CI we have no rights to drop and create DB.
In the tests we are downgrading and upgrading DB schema by Alembic
migrations. In case of adding migrations in one reveiw we had
all other reviews broken due to Alembic failures on attempt to
downgrade DB schema.
As fix we are fallback to the clening DB without migrations
in case of Alembic error on the schema downgrade.

Change-Id: I65faf43c28427594495c67696ca190e7fe29f083
Closes-Bug: #1556876