PoC for data versioning of installation info
Tables for saving resource changes are introduced.
In table 'history' we save JSON representation of diff, that should
be applied to the original data to restore the previous object state.

For diff calculation we are introducing 'history_last_snapshot' table.
In this table we are store the full snapshot of tracked resource.

For restoring resource state diffs from 'history' table should be
applied to the 'last snapshot' from newest to oldest.

Data versioning imlementation is abstract and can be applied for
any resource type. Here we are using it for tracking installation
info changes.

Change-Id: I5406be1b68f561d90ff0ca11e8e0ac7ab45096ee