Connection hanging issue fixed
On huge reports data we can have 500 HTTP error by timeout limitation.
In case of streaming data from the DB cursor we had hanged
DB operation inside the open transaction.

Releasing SQLAlchemy session was added to collector and analytics
apps.

We are migrating ActionLogs content to the Elasticsearch, but that
data is not used yet for any report. We are stop to mirate this
data to Elasticearch.

Change-Id: I009081a2304b36fb4ab3705b6fa09a8a3df5710b
Closes-Bug: #1564427